improve interface, add information about the feature, and fix paused timers
